requirements.txt
torch==2.2.2
torchaudio==2.2.2
transformers==4.39.0
numpy>=1.21.0
tqdm>=4.65.0
scipy>=1.7.0
scikit-learn>=1.0.0
soundfile>=0.12.1
SpeechRecognition>=3.10.0
pydub>=0.25.1
langchain>=0.1.0
langchain-community>=0.0.15
sentence-transformers>=2.2.2
chromadb>=0.4.22
biopython>=1.81
librosa>=0.10.0
tensorflow>=2.12.0
keras>=2.12.0
audiocraft>=1.0.0
streamlit>=1.24.0
spacy>=3.5.0
torchmetrics>=0.11.0
torchvision>=0.14.0
sentencepiece>=0.1.99
lxml>=4.9.0
beautifulsoup4>=4.11.0  
requests>=2.28.0
google-search-results>=2.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0